Senior Research Project "The Helping Hand"
The aim of this project is to make a website with similar to those of
gofundme/indigogo/kickstarter with the focus of helping those in need.
While providing the tools necessary for both the one in need and the
benifactor(s) to keep in contact, with an emphasis on the one receiving
money to post photographic or status updates of how the money is being
put to use to help. Thus providing the financial backers some
reassurement that their generosity is not being taken for granted and
will hopefully bolster more charitable acts from them and their peers.